wip: trying to get filenames figured out... really messed up right now

+++ app.py
@@ -73,13 +73,19 @@ def inference_main(
     print("done")
     # eventually use diretories instead of mutating song names - should make a database easier
     # mp3 conversion
-    instrumental_output_file = f"/app-data/instrumentals/{basename}.wav"
-    instrumental_output_mp3 = f"/app-data/instrumentals/{basename}.mp3"
-    sf.write(instrumental_output_file, wave.T, sr)
+
+    (
+        instrumental_filename_wav,
+        instrumental_filename,
+        vocal_filename_wav,
+        vocal_filename,
+    ) = get_filenames(basename, tta, postprocess)
+
+    sf.write(instrumental_filename_wav, wave.T, sr)
 
     print("converting to mp3...", end=" ")
-    sound = pydub.AudioSegment.from_wav(instrumental_output_file)
-    sound.export(instrumental_output_mp3, format="mp3")
+    sound = pydub.AudioSegment.from_wav(instrumental_filename_wav)
+    sound.export(instrumental_filename, format="mp3")
     print("done")
 
     print("inverse stft of vocals...", end=" ")
@@ -87,17 +93,37 @@ def inference_main(
     print("done")
     # eventually use diretories instead of mutating song names - should make a database easier
     # mp3 conversion
-    vocal_output_file = f"/app-data/vocals/{basename}.wav"
-    vocal_output_mp3 = f"/app-data/vocals/{basename}.mp3"
 
-    sf.write(vocal_output_file, wave.T, sr)
+    sf.write(vocal_filename_wav, wave.T, sr)
 
     print("converting to mp3...", end=" ")
-    sound = pydub.AudioSegment.from_wav(vocal_output_file)
-    sound.export(vocal_output_mp3, format="mp3")
+    sound = pydub.AudioSegment.from_wav(vocal_filename_wav)
+    sound.export(vocal_filename, format="mp3")
     print("done")
 
-    return instrumental_output_mp3, vocal_output_mp3
+    return None
+
+
+def get_filenames(basename, use_tta, use_postprocess):
+
+    suffix = ""
+    if use_tta:
+        suffix += ".tta"
+    if use_postprocess:
+        suffix += ".postprocess"
+
+    instrumental_filename = "/app-data/instrumentals/" + basename + f"{suffix}.mp3"
+    instrumental_filename_wav = "/app-data/instrumentals/" + basename + f"{suffix}.wav"
+
+    vocal_filename = "/app-data/vocals/" + basename + f"{suffix}.mp3"
+    vocal_filename_wav = "/app-data/vocals/" + basename + f"{suffix}.wav"
+
+    return (
+        instrumental_filename_wav,
+        instrumental_filename,
+        vocal_filename_wav,
+        vocal_filename,
+    )
 
 
 @contextmanager
@@ -134,9 +160,36 @@ def st_stderr(dst):
         yield
 
 
+def check_if_already_processed(instrumental_filename, vocal_filename):
+    if Path(instrumental_filename).exists() and Path(vocal_filename).exists():
+        return True
+    return False
+
+
+def check_and_download(
+    instrumental_filename,
+    vocal_filename,
+):
+    try:
+        with open(instrumental_filename, "rb") as f:
+            st.download_button(
+                "Download Instrumentals",
+                f,
+                file_name=instrumental_filename,
+            )
+        with open(vocal_filename, "rb") as f:
+            st.download_button(
+                "Download Vocals",
+                f,
+                file_name=vocal_filename,
+            )
+    except FileNotFoundError:
+        raise FileNotFoundError
+
+
 def main():
-    instrumental_output_mp3 = ""
-    vocal_output_mp3 = ""
+    # instrumental_output_mp3 = ""
+    # vocal_output_mp3 = ""
 
     with st.form("song_form"):
         data = st.file_uploader("Upload mp3 to split instrumentals from", type=["mp3"])
@@ -148,29 +201,44 @@ def main():
         )
 
         submitted = st.form_submit_button("Punch it Chewy!")
+        basename = data.name
+        (
+            instrumental_filename_wav,
+            instrumental_filename,
+            vocal_filename_wav,
+            vocal_filename,
+        ) = get_filenames(basename, use_tta, use_postprocess)
+
         if submitted:
-            with st_stdout("success"):
-                print("Reading Data\n")
-                audio_bytes = data.read()
+            if check_if_already_processed(instrumental_filename, vocal_filename):
+                with st_stdout("success"):
+                    print("File already processed")
+            else:
 
-                # write out to file until I figure out how to just pass audio_bytes to the inference script
-                print("Saving to staging file")
-                with open(f"/app-data/{data.name}", "wb") as outfile:
-                    outfile.write(audio_bytes)
-                # allow listening to song through app as sanity check
-                # st.audio(audio_bytes, format="audio/mp3")
-            with st_stdout("code"):
-                instrumental_output_mp3, vocal_output_mp3 = inference_main(
-                    f"/app-data/{data.name}"
-                )
+                with st_stdout("error"):
+                    print("Still waiting...")
 
-            with open(instrumental_output_mp3, "rb") as f:
+                with st_stdout("error"):
+                    with st_stdout("success"):
+                        print("Reading Data\n")
+                        audio_bytes = data.read()
+
+                        # write out to file until I figure out how to just pass audio_bytes to the inference script
+                        print("Saving to staging file")
+                        with open(f"/app-data/{data.name}", "wb") as outfile:
+                            outfile.write(audio_bytes)
+                        # allow listening to song through app as sanity check
+                        # st.audio(audio_bytes, format="audio/mp3")
+                with st_stdout("code"):
+                    inference_main(f"/app-data/{data.name}")
+
+            with open(instrumental_filename, "rb") as f:
                 instrumental_output_bytes = f.read()
             # allow listening to song through app as sanity check
             st.write("Have a listen to the instrumental!")
             st.audio(instrumental_output_bytes, format="audio/mp3")
 
-            with open(vocal_output_mp3, "rb") as f:
+            with open(vocal_filename, "rb") as f:
                 vocal_output_bytes = f.read()
             # allow listening to song through app as sanity check
             st.write("Have a listen to the vocals!")
@@ -179,36 +247,19 @@ def main():
     # TODO: download it...
     st.markdown("### Once the song is split you will be able to download below")
     try:
-        suffix = "-Instrumental"
-        if use_tta:
-            suffix += ".tta"
-        if use_postprocess:
-            suffix += ".postprocess"
-        with open(instrumental_output_mp3, "rb") as f:
-            st.download_button(
-                "Download Instrumentals",
-                f,
-                file_name=Path(instrumental_output_mp3).stem + f"{suffix}.mp3",
-            )
-        suffix = "-Vocals"
-        if use_tta:
-            suffix += ".tta"
-        if use_postprocess:
-            suffix += ".postprocess"
-        with open(vocal_output_mp3, "rb") as f:
-            st.download_button(
-                "Download Vocals",
-                f,
-                file_name=Path(vocal_output_mp3).stem + f"{suffix}.mp3",
-            )
+        check_and_download(instrumental_filename, vocal_filename)
+
     except FileNotFoundError:
         with st_stdout("error"):
             print("Still waiting...")
 
 
-if __name__ == "__main__":
-    st.markdown("# Audacity who??")
-    st.markdown(
-        "**Upload one mp3 at a time and the vocal-remover app will strip the instrumentals from the vocals. There are 2 options to play with and you can listen to the instrumental/vocal tracks separately from here before downloading**"
-    )
-    main()
+st.markdown("# Audacity who??")
+st.markdown(
+    "**Upload one mp3 at a time and the vocal-remover app will strip the instrumentals from the vocals. There are 2 options to play with and you can listen to the instrumental/vocal tracks separately from here before downloading**"
+)
+main()
+
+# TODO: download it...
+# st.markdown("### Once the song is split you will be able to download below")
+# check_and_download(use_tta, use_postprocess, instrumental_output_mp3, vocal_output_mp3)
